titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/util-adaptors/src/lib/helpers/stock-receive-stock.helper.ts
import { StockReceiveStock } from '@poss-web/shared/models';
import { StockReceiveStockAdaptor } from '../stock-receive/stock-receive-stock.adaptor';

export class StockReceiveStockHelper {
  static getStocks(
    data: any,
    typeField: string
  ): { stocks: StockReceiveStock[]; count: number } {
    const stocks: StockReceiveStock[] = [];
    for (const stock of data.results) {
      stocks.push(StockReceiveStockAdaptor.fromJson(stock, typeField));
    }
    return { stocks, count: data.totalElements };
  }
}
